nudb: Move posix_file, win32_file to public directory
